+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 
 * Update README.
 * Stable release code.
+
+## [0.1.1] - 01-06-2019
+
+* Add speed and opacity controls
+* Assign types to all parameters
+* Assign a defautl value to DotSize - DotSize.random